split pie trace base plot module from cartesian:
- so that _has??? are one-to-one with base plot modules
- handle case where base plot module attrRegex isn't defined
  in plot schema.
